Clean up the files and names
